Refactor compareChoices and result functions
compareChoices originally had a nested switch structure to handle all 9 possible combinations of player and computer choices. Now compareChoices first handles ties by comparing player and computer. If it is not a tie, the nested switch structure handles the remaining 6 combinations of choices.

Additionally, compareChoices used to run the result functions; win, lose, and tie. Instead compareChoices now returns the function to be called within playRound. The result functions print a message to console and return a result string to be returned by playRound.