Refactor incrementValue function
As mentioned in previous commit, I decided to refactor the logic such that all functions incrementing a textContent by 1 value call incrementValue as opposed to different functions. incrementValue takes a string as an argument that is passed by the calling function to determine which element is being incremented and how the value is parsed from the textContent.